hospitals/requirements.txt
Flask==2.1.2
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
pyjwt==2.1.0
Werkzeug==2.1.2
Flask-SQLAlchemy==2.5.1

